+++ UltraStock/Controllers/UsuarioController.cs
@@ -2,6 +2,7 @@
 using Microsoft.EntityFrameworkCore;
 using System.Collections.Generic;
 using UltraStock.Data;
+using UltraStock.Helpers;
 using UltraStock.Models;
 
 namespace UltraStock.Controllers
@@ -9,7 +10,6 @@ namespace UltraStock.Controllers
     public class UsuarioController : Controller
     {
         private readonly AppDbContext _context;
-
         public UsuarioController(AppDbContext context)
         {
             _context = context;
@@ -17,6 +17,11 @@ namespace UltraStock.Controllers
 
         public IActionResult Index()
         {
+            if (HttpContext.Session.GetString("Usuario") == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
             var usuarios = _context.Usuarios.ToList();
 
             return View(usuarios);
@@ -26,29 +31,79 @@ namespace UltraStock.Controllers
 
         public IActionResult Create()
         {
+            if (HttpContext.Session.GetString("Usuario") == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
             return View();
         }
 
         [HttpPost]
         public IActionResult Create(Usuario usuario)
         {
-            _context.Usuarios.Add(usuario);
-            _context.SaveChanges();
+            if (HttpContext.Session.GetString("Usuario") == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
 
-            return RedirectToAction("Index");
+            if (ModelState.IsValid)
+            {
+                //Convertir contraseña a Hash
+                usuario.Clave = HashHelper.ObtenerHash(usuario.Clave);
+                _context.Usuarios.Add(usuario);
+                _context.SaveChanges();
+
+                return RedirectToAction("Index");
+            }
+
+            return View(usuario);
         }
 
         //Formulario editar
         public IActionResult Edit(int id)
         {
+            if (HttpContext.Session.GetString("Usuario") == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
             var usuario = _context.Usuarios.Find(id);
+            ViewBag.Usuarios = _context.Usuarios.ToList();
+
 
             return View(usuario);
+        }
+
+        //Actualizar producto
+        [HttpPost]
+        public IActionResult Edit(Usuario usuario)
+        {
+            if (HttpContext.Session.GetString("Usuario") == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
+            _context.Usuarios.Update(usuario);
+            _context.SaveChanges();
+
+            return RedirectToAction("Index");
         }
 
         //Eliminar usuario
         public IActionResult Delete(int id)
         {
+            if (HttpContext.Session.GetString("Usuario") == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
+            var rol = HttpContext.Session.GetString("Rol");//Solo admin puede eliminar
+            if (rol != "admin")
+            {
+                return RedirectToAction("Index");
+            }
+
             var usuario = _context.Usuarios.Find(id);
 
             _context.Usuarios.Remove(usuario);
